Re-design transfer rules!
The tables had to change to accomodate the "source discipline" actually being
an arbitrary string (component subject area). This is important because rules
can cover source courses from multiple disciplines.

This is a work in progress. So far, the tables have been updated and the
rule_groups table has been renamed to transfer_rules. The populate_rule_groups
script has been renamed to populae_transfer_rules.

In an attempt to speed up the re-build process, a 2-minute pass builds the bones
of the three tables in memory using a single pass over the .csv query output
from CF. Next will be to look up all course_ids to verify that they are not
bogus (and delete any rules where they are), to create a list of all the source
disciplines covered by eeach rule (for map_courses/review_rules), and to clone
rules for all cross-listed courses in the source and destination lists for each
rule so all the courses will show up in the map and review parts of the app.
Then the dicts will be used to build the db tables.

Sigh.