New transfer tables ready for testing.
transfer_rules, source_courses, and destination_courses.
Still takes forever to build the tables.

And I still don't understand why all three tables have the same length. The
the point of separating them was because the number of source courses and
destination courses might be unequal for a single rule.