New structures for source and destination courses.
Instead of the rule key (which was redundant to the rule_id), they now have the
course discipline and cuny_subject for more efficient targeted searches.

By keeping a cache of the courses table in memory and reducing the number of
steps from 4 to 2, populate_transfer_rule.py now completes it's job about 40%
faster than before, and with more detailed anomaly reporting.