Add effective date to transfer_rules.
The CF query was modified to return the effective dates from each of the tables
that go into the query. Now populate_transfer_rules.py takes the max of each
as the effective date of a rule part (row of the CF query) and the max of all
rule parts as the effective date of the rule.

The rationale is that OUR decided that new rules should not take effect during a
semester, so they future-date changed/new rules to the end of the current
semester. By adding the effective date to our transfer_rules table, we can mark
those rules (somehow) so people will understand the difference between now and
then.